fix cli

+++ linalg/cost_model/src/main.rs
@@ -526,7 +526,7 @@ fn main() {
                 .arg(
                     Arg::new("size")
                     .short('s')
-                    .short("size")
+                    .long("size")
                     .help("Sample size (total)")
                     .takes_value(true)
                     .default_value("128"),
